Minor tweaks

+++ WinApp/PlayPauser/Parts/NoSleepPart.cs
@@ -1,30 +1,23 @@
-﻿using System;
-using System.Diagnostics;
+﻿using System.Diagnostics;
 using System.IO;
 using System.Reflection;
-using System.Runtime.InteropServices;
 
 namespace PlayPauser.Parts
 {
     public class NoSleepPart : IPart
     {
-        [DllImport("kernel32.dll", CharSet = CharSet.Auto, SetLastError = true)]
-        private static extern ExecutionState SetThreadExecutionState(ExecutionState esFlags);
-
-        [Flags]
-        private enum ExecutionState : uint
-        {
-            EsAwaymodeRequired = 0x00000040,
-            EsContinuous = 0x80000000,
-            EsDisplayRequired = 0x00000002,
-            EsSystemRequired = 0x00000001
-        }
-
+        private bool enabled = true;
         private string fileName;
         private Process p;
 
         public void Start(Options options)
         {
+            enabled = options.IsNoSleepEnabled;
+            if (!enabled)
+            {
+                return;
+            }
+
             fileName = Path.Combine(Path.GetTempPath(), "NoSleep.exe");
             if (File.Exists(fileName))
             {
@@ -36,12 +29,17 @@ namespace PlayPauser.Parts
             p = Process.Start(new ProcessStartInfo
             {
                 FileName = fileName,
-                WindowStyle = ProcessWindowStyle.Hidden
+                WindowStyle = ProcessWindowStyle.Hidden,
             });
         }
 
         public void Stop()
         {
+            if (!enabled)
+            {
+                return;
+            }
+
             p.Kill();
             p.WaitForExit();
